Cybersec/Libraries/whois_with_pdf/requirements.txt
fpdf         1.7.2
python-whois 0.7.3
pip-autoremove 0.10.0
pipdeptree     2.2.1
